making buildFid art.cc internal only.. moving frameid enums to art_defs.h
