+++ src/art.h
@@ -1,7 +1,6 @@
 #ifndef ART_H
 #define ART_H
 
-#include <assert.h>
 #include <memory>
 
 #include "animation.h"
@@ -39,47 +38,6 @@ extern CritterFrameId _art_vault_person_nums[DUDE_NATIVE_LOOK_COUNT][GENDER_COUN
 
 extern Cache gArtCache;
 
-int artInit();
-void artReset();
-void artExit();
-char* artGetObjectTypeName(ObjectType objectType);
-int artIsObjectTypeHidden(ObjectType objectType);
-void artToggleObjectTypeHidden(ObjectType objectType);
-int artGetFidgetCount(int headFid);
-void artRender(int fid, unsigned char* dest, int width, int height, int pitch);
-int art_list_str(int fid, char* name);
-Art* artLock(int fid, CacheEntry** cache_entry);
-unsigned char* artLockFrameData(int fid, int frame, Rotation rotation, CacheEntry** out_cache_entry);
-int artUnlock(CacheEntry* cache_entry);
-int artCacheFlush();
-int artCopyFileName(ObjectType objectType, int id, char* dest);
-int _art_get_code(AnimationType animation, WeaponAnimation weaponType, char* weaponCodePtr, char* animationCodePtr);
-char* artBuildFilePath(int fid);
-int artGetFramesPerSecond(Art* art);
-int artGetActionFrame(Art* art);
-int artGetFrameCount(Art* art);
-int artGetWidth(Art* art, int frame = 0, Rotation rotation = ROTATION_NE);
-int artGetHeight(Art* art, int frame = 0, Rotation rotation = ROTATION_NE);
-int artGetSize(Art* art, int frame, Rotation rotation, int* out_width, int* out_height);
-int artGetFrameOffsets(const Art* art, int frame, Rotation rotation, int* xPtr, int* yPtr);
-int artGetRotationOffsets(Art* art, Rotation rotation, int* out_offset_x, int* out_offset_y);
-unsigned char* artGetFrameData(Art* art, int frame = 0, Rotation rotation = ROTATION_NE);
-unsigned char* artGetFrameData(const Art* art, int frame, Rotation rotation, int* widthPtr, int* heightPtr, int* xOffsetPtr, int* yOffsetPtr);
-ArtFrame* artGetFrame(const Art* art, int frame, Rotation rotation);
-ConstBuffer2D artGetFrameBuffer(const Art* art, int frame, Rotation rotation);
-bool artExists(int fid);
-bool _art_fid_valid(int fid);
-CritterFrameId _art_alias_num(CritterFrameId index);
-int artCritterFidShouldRun(int fid);
-int artAliasFid(int fid);
-int buildFid(ObjectType objectType, int frmId, int animType = 0, int weaponCode = 0, Rotation rotation = ROTATION_NE);
-int artListIndex(ObjectType objectType, const char* name);
-Art* artLoad(const char* path);
-int artRead(const char* path, unsigned char* data);
-int artWrite(const char* path, unsigned char* data);
-
-using ArtPtr = InternalPtr<Art>;
-
 class NamedCacheEntry;
 std::shared_ptr<NamedCacheEntry> artLockNamedFrameData(const char* path);
 
@@ -106,74 +64,19 @@ public:
     {
     }
 
-    explicit FrmId(MiscFrameId misc, AnimationType animType = ANIM_STAND)
-        : _objectType(OBJ_TYPE_MISC)
-        , _fid(buildFid(OBJ_TYPE_MISC, misc, animType))
-    {
-    }
-
-    explicit FrmId(SceneryFrameId scenery)
-        : _objectType(OBJ_TYPE_SCENERY)
-        , _fid(buildFid(OBJ_TYPE_SCENERY, scenery))
-    {
-    }
-
-    explicit FrmId(WallFrameId wall)
-        : _objectType(OBJ_TYPE_WALL)
-        , _fid(buildFid(OBJ_TYPE_WALL, wall))
-    {
-    }
-
-    explicit FrmId(ItemFrameId item)
-        : _objectType(OBJ_TYPE_ITEM)
-        , _fid(buildFid(OBJ_TYPE_ITEM, item))
-    {
-    }
-
-    explicit FrmId(TileFrameId tile)
-        : _objectType(OBJ_TYPE_TILE)
-        , _fid(buildFid(OBJ_TYPE_TILE, tile))
-    {
-    }
-
-    explicit FrmId(SkillDexFrameId skilldex)
-        : _objectType(OBJ_TYPE_SKILLDEX)
-        , _fid(buildFid(OBJ_TYPE_SKILLDEX, skilldex))
-    {
-    }
-
-    explicit FrmId(InterfaceFrameId interface)
-        : _objectType(OBJ_TYPE_INTERFACE)
-        , _fid(buildFid(OBJ_TYPE_INTERFACE, interface))
-    {
-    }
-
-    explicit FrmId(CritterFrameId critter, AnimationType animType, WeaponAnimation weaponAnimation, Rotation rotation)
-        : _objectType(OBJ_TYPE_CRITTER)
-        , _fid(buildFid(OBJ_TYPE_CRITTER, critter, animType, weaponAnimation, rotation))
-    {
-    }
-
-    explicit FrmId(Object* object, AnimationType animType, WeaponAnimation weaponAnimation, Rotation rotation)
-        : _objectType(objectTypeFromFid(object->fid))
-        , _fid(object == nullptr ? -1 : buildFid(objectTypeFromFid(object->fid), objectFrameIdFromFid(object->fid), animType, weaponAnimation, rotation))
-    {
-    }
-
-    explicit FrmId(Head head, HeadAnimation headAnimation = HEAD_ANIMATION_VERY_GOOD_REACTION, int fidget = 0)
-        : _objectType(OBJ_TYPE_HEAD)
-        , _fid(buildFid(OBJ_TYPE_HEAD, head, headAnimation, fidget))
-    {
-    }
-
-    explicit FrmId(Background background)
-        : _objectType(OBJ_TYPE_BACKGROUND)
-        , _fid(buildFid(OBJ_TYPE_BACKGROUND, background))
-    {
-    }
-
+    explicit FrmId(MiscFrameId misc, AnimationType animType = ANIM_STAND);
+    explicit FrmId(SceneryFrameId scenery);
+    explicit FrmId(WallFrameId wall);
+    explicit FrmId(ItemFrameId item);
+    explicit FrmId(TileFrameId tile);
+    explicit FrmId(SkillDexFrameId skilldex);
+    explicit FrmId(InterfaceFrameId interface);
+    explicit FrmId(CritterFrameId critter, AnimationType animType, WeaponAnimation weaponAnimation, Rotation rotation);
+    explicit FrmId(Object* object, AnimationType animType, WeaponAnimation weaponAnimation, Rotation rotation);
+    explicit FrmId(HeadFrameId head, HeadAnimation headAnimation = HEAD_ANIMATION_VERY_GOOD_REACTION, int fidget = 0);
+    explicit FrmId(BackgroundFrameId background);
     explicit FrmId(ObjectType objType, const char* path);
-    explicit FrmId(const char* path);
+    explicit FrmId(ObjectType objectType, int frmId, int animType = 0, int weaponCode = 0, Rotation rotation = ROTATION_NE);
 
     int fid() const { return _fid; }
     bool hasObjectType() const { return objectTypeIsValid(_objectType); }
@@ -195,6 +98,60 @@ private:
     ObjectType _objectType = OBJ_TYPE_INVALID;
     const char* _path = nullptr;
 };
+
+int artInit();
+void artReset();
+void artExit();
+char* artGetObjectTypeName(ObjectType objectType);
+int artIsObjectTypeHidden(ObjectType objectType);
+void artToggleObjectTypeHidden(ObjectType objectType);
+int artGetFidgetCount(int headFid);
+void artRender(int fid, unsigned char* dest, int width, int height, int pitch);
+int art_list_str(int fid, char* name);
+Art* artLock(int fid, CacheEntry** cache_entry);
+
+inline Art* artLock(const FrmId& frmId, CacheEntry** cache_entry)
+{
+    return artLock(frmId.fid(), cache_entry);
+}
+
+unsigned char* artLockFrameData(int fid, int frame, Rotation rotation, CacheEntry** out_cache_entry);
+int artUnlock(CacheEntry* cache_entry);
+int artCacheFlush();
+int artCopyFileName(ObjectType objectType, int id, char* dest);
+int _art_get_code(AnimationType animation, WeaponAnimation weaponType, char* weaponCodePtr, char* animationCodePtr);
+char* artBuildFilePath(int fid);
+int artGetFramesPerSecond(Art* art);
+int artGetActionFrame(Art* art);
+int artGetFrameCount(Art* art);
+int artGetWidth(Art* art, int frame = 0, Rotation rotation = ROTATION_NE);
+int artGetHeight(Art* art, int frame = 0, Rotation rotation = ROTATION_NE);
+int artGetSize(Art* art, int frame, Rotation rotation, int* out_width, int* out_height);
+int artGetFrameOffsets(const Art* art, int frame, Rotation rotation, int* xPtr, int* yPtr);
+int artGetRotationOffsets(Art* art, Rotation rotation, int* out_offset_x, int* out_offset_y);
+unsigned char* artGetFrameData(Art* art, int frame = 0, Rotation rotation = ROTATION_NE);
+unsigned char* artGetFrameData(const Art* art, int frame, Rotation rotation, int* widthPtr, int* heightPtr, int* xOffsetPtr, int* yOffsetPtr);
+ArtFrame* artGetFrame(const Art* art, int frame, Rotation rotation);
+ConstBuffer2D artGetFrameBuffer(const Art* art, int frame, Rotation rotation);
+bool artExists(int fid);
+
+inline bool artExists(const FrmId& frmId) {
+    return artExists(frmId.fid());
+}
+
+bool _art_fid_valid(int fid);
+CritterFrameId _art_alias_num(CritterFrameId index);
+int artCritterFidShouldRun(int fid);
+int artAliasFid(int fid);
+int artListIndex(ObjectType objectType, const char* name);
+Art* artLoad(const char* path);
+int artRead(const char* path, unsigned char* data);
+int artWrite(const char* path, unsigned char* data);
+
+using ArtPtr = InternalPtr<Art>;
+
+class NamedCacheEntry;
+std::shared_ptr<NamedCacheEntry> artLockNamedFrameData(const char* path);
 
 // RAII helper for locking one selected frame from FID-backed or path-backed art.
 // lock/unlock use caches instead of just loading/unloading directly.
@@ -242,15 +199,6 @@ private:
     int _xOffset = 0;
     int _yOffset = 0;
 };
-
-inline bool artExists(const FrmId& frmId) {
-    return artExists(frmId.fid());
-}
-
-inline Art* artLock(const FrmId& frmId, CacheEntry** cache_entry)
-{
-    return artLock(frmId.fid(), cache_entry);
-}
 
 } // namespace fallout
